Added input options to useKmenu hook #5
